feat: improve audio buffer (see desc.)
The audio buffer had 2 issues
- The buffer would be empty until it was given some data by the read_ahead task, which the player would interpret as no more audio, and exit
- The buffer could have a bit of data left, if the audio data didnt perfectly divide into the frame length

This fixes both. AudioBuffer no longer waits for a complete frame, instead it pads its output to be the correct length. Note this only occurs at the end of a file
AudioBuffer now has an event which fires when the buffer has been *initialised* (received its first data). The audio object waits for this event when creating processes